Move files up one directory from ‘src’
