+++ test/off.test.ts
@@ -2,10 +2,10 @@
 // Copyright (c) roydukkey. All rights reserved.                     //
 // ================================================================= //
 
-import { store } from '../../src/support/store';
-import { off as mainOff, on as mainOn } from '../../dist/holosun';
+import { store } from '../src/support/store';
+import { off as mainOff, on as mainOn } from '../dist/holosun';
 import { off as moduleOff, on as moduleOn } from 'holosun';
-import { off as srcOff, on as srcOn } from '../../typescript';
+import { off as srcOff, on as srcOn } from '../typescript';
 
 
 const packages: Array<[string, typeof srcOff, typeof srcOn, typeof store | null]> = [
